[REFACTOR] rename atchange to attendancechange
